requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
pluggy==1.0.0
    # via -r requirements.in
robotframework==4.1.3
    # via -r requirements.in
